Almost Done

+++ backend/listings/models.py
@@ -29,7 +29,7 @@ class Listing(models.Model):
                                  default=SaleTypes.FOR_SALE)
     price = models.IntegerField()
     bedrooms = models.IntegerField()
-    bathrooms = models.DecimalField(max_digits=2, decimal_places=1)
+    bathrooms = models.IntegerField()
     home_type = models.CharField(max_length=50,
                                  choices=HomeTypes.choices,
                                  default=HomeTypes.HOUSE)
